Feat & Refactor: 로딩 표시 추가, 마크업/문서 링크 로직 리펙토링

+++ src/App.js
@@ -4,6 +4,7 @@ import router from "./router.js";
 import { asyncDataObj, request } from "./api.js";
 import { setItem, getItem, removeItem } from "./storage.js";
 import { getLocalSaveKey } from "./utility.js";
+import Indicator from "./Indicator.js";
 
 export default function App({ targetEl }) {
   this.isInit = false;
@@ -14,7 +15,7 @@ export default function App({ targetEl }) {
     document: { ...asyncDataObj },
   };
 
-  this.setState = (nextState) => {
+  this.setState = (nextState) => { console.log(nextState)
     const prevState = JSON.parse(JSON.stringify(this.state));
 
     if (JSON.stringify(prevState) !== JSON.stringify(nextState)) {
@@ -36,15 +37,20 @@ export default function App({ targetEl }) {
         JSON.stringify(prevState.selectedDocumentId) !==
           JSON.stringify(nextState.selectedDocumentId) ||
         JSON.stringify(prevState.document) !==
-          JSON.stringify(nextState.document) ||
-        JSON.stringify(prevState.flatDocuments) !==
-          JSON.stringify(nextState.flatDocuments)
+          JSON.stringify(nextState.document)
       ) {
-        editor.setState({
-          selectedDocumentId: this.state.selectedDocumentId,
-          document: this.state.document,
-          flatDocuments: this.state.flatDocuments,
-        });
+        editor.setState({ ...this.state });
+      }
+
+      if (
+        JSON.stringify(prevState.document) !==
+          JSON.stringify(nextState.document) ||
+        JSON.stringify(prevState.documents) !==
+          JSON.stringify(nextState.documents)
+      ) {
+        indicator.setState(
+          this.state.document.isLoading || this.state.documents.isLoading
+        );
       }
 
       this.render();
@@ -54,6 +60,11 @@ export default function App({ targetEl }) {
   let serverUpdateTimer = null;
   let localSaveTimer = null;
   let optimisticUpdateTimer = null;
+
+  const indicator = new Indicator({
+    targetEl,
+    initialState: false,
+  });
 
   const sidebar = new Sidebar({
     targetEl,
@@ -108,33 +119,6 @@ export default function App({ targetEl }) {
     },
   });
 
-  const updateFlatDocuments = async () => {
-    const { isLoading, isError, data } = this.state.documents;
-
-    if (isLoading || isError || !data) {
-      return null;
-    }
-
-    const flatDocuments = [];
-
-    function recursion(documents) {
-      for (const document of documents) {
-        flatDocuments.push({ id: document.id, title: document.title });
-
-        if (document.documents && document.documents.length > 0) {
-          recursion(document.documents);
-        }
-      }
-    }
-
-    recursion(this.state.documents.data);
-
-    this.setState({
-      ...this.state,
-      flatDocuments,
-    });
-  };
-
   const optimisticUpdate = async ({ id, title }) => {
     const newData = JSON.parse(JSON.stringify(this.state.documents.data));
 
@@ -166,11 +150,11 @@ export default function App({ targetEl }) {
   const fetchDocuments = async () => {
     try {
       const res = await request("/documents");
+
       this.setState({
         ...this.state,
         documents: { ...asyncDataObj, data: res },
       });
-      updateFlatDocuments();
     } catch (e) {
       this.setState({
         ...this.state,
@@ -185,7 +169,15 @@ export default function App({ targetEl }) {
         ...this.state,
         document: { ...asyncDataObj },
       });
+
       return null;
+    }
+
+    if (id !== this.state.document.data?.id) {
+      this.setState({
+        ...this.state,
+        document: { ...asyncDataObj, isLoading: true },
+      });
     }
 
     const LOCAL_SAVE_KEY = getLocalSaveKey(id);
@@ -253,9 +245,9 @@ export default function App({ targetEl }) {
   };
 
   const deleteDocument = async (id) => {
-    // if (!id || typeof id !== "number") {
-    //   throw new Error("삭제할 문서의 id를 입력하세요!");
-    // }
+    if (!id || typeof id !== "number") {
+      throw new Error("삭제할 문서의 id를 입력하세요!");
+    }
 
     const res = await request(`/documents/${id}`, { method: "DELETE" });
 
@@ -295,6 +287,7 @@ export default function App({ targetEl }) {
       router.init(onRouterChange);
       onRouterChange(window.location.pathname);
 
+      indicator.render();
       sidebar.render();
       editor.render();
